# okito-server/src/main/resources/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "PostComment": {
        "properties": {
          "id": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "290"
          },
          "content": {
            "format": "string",
            "description": "No comments found.",
            "type": "string",
            "example": "hogxum"
          },
          "uploadTime": {
            "format": "string",
            "description": "No comments found.",
            "type": "string",
            "example": "2023-02-24 11:14:15"
          },
          "authorId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "90"
          },
          "postId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "752"
          },
          "parentId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "69"
          }
        }
      },
...
      "RespResult": {
        "properties": {
          "timestamp": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "1677208455247"
          },
          "status": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "107"
          },
          "message": {
            "format": "string",
            "description": "No comments found.",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "No comments found.(object)",
            "type": "object",
            "example": ""
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8088/"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
...
    "/post/comment/count/post/{postId}": {
      "get": {
        "summary": "count postComments By post id",
        "deprecated": false,
        "description": "count postComments By post id",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 740,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentcountpost{postId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "post id",
            "required": true,
            "example": "233",
            "name": "postId"
          }
        ]
      }
    },
...
    "/post/comment": {
      "post": {
        "summary": "add a postComment",
        "deprecated": false,
        "description": "add a postComment",
        "tags": [
          "post comment controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostComment"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": 729,\n  \"content\": \"mwycsz\",\n  \"uploadTime\": \"2023-02-24 11:14:15\",\n  \"authorId\": 65,\n  \"postId\": 925,\n  \"parentId\": 829\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 648,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcomment",
        "parameters": []
      },
      "get": {
        "summary": "select all postComments",
        "deprecated": false,
        "description": "select all postComments",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 151,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcomment",
        "parameters": []
      }
    },
    "/pin/comment": {
      "post": {
        "summary": "add a pinComment",
        "deprecated": false,
        "description": "add a pinComment",
        "tags": [
          "pin comment controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinComment"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": 136,\n  \"content\": \"rmqewv\",\n  \"uploadTime\": {},\n  \"authorId\": 392,\n  \"pinId\": 851,\n  \"parentId\": 90\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 913,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincomment",
        "parameters": []
      },
      "get": {
        "summary": "select all pinComments",
        "deprecated": false,
        "description": "select all pinComments",
        "tags": [
          "pin comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 739,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincomment",
        "parameters": []
      }
    },
...
    "/post/comment/{id}": {
      "delete": {
        "summary": "delete pinComment By id",
        "deprecated": false,
        "description": "delete pinComment By id",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 686,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcomment{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "id",
            "required": true,
            "example": "598",
            "name": "id"
          }
        ]
      },
      "get": {
        "summary": "select postComment By id",
        "deprecated": false,
        "description": "select postComment By id",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 251,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcomment{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "id",
            "required": true,
            "example": "563",
            "name": "id"
          }
        ]
      }
    },
    "/pin/comment/{id}": {
      "delete": {
        "summary": "delete pinComment By id",
        "deprecated": false,
        "description": "delete pinComment By id",
        "tags": [
          "pin comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 295,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincomment{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "id",
            "required": true,
            "example": "878",
            "name": "id"
          }
        ]
      },
      "get": {
        "summary": "select pinComment By id",
        "deprecated": false,
        "description": "select pinComment By id",
        "tags": [
          "pin comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 825,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincomment{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "id",
            "required": true,
            "example": "820",
            "name": "id"
          }
        ]
      }
    },
...
    "/pin/comment/author/{authorId}": {
      "get": {
        "summary": "select all pinComments by authorId",
        "deprecated": false,
        "description": "select all pinComments by authorId",
        "tags": [
          "pin comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 87,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincommentauthor{authorId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "authorId",
            "required": true,
            "example": "759",
            "name": "authorId"
          }
        ]
      }
    },
    "/post/comment/author/{authorId}": {
      "get": {
        "summary": "select all postComments by authorId",
        "deprecated": false,
        "description": "select all postComments by authorId",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 210,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentauthor{authorId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "authorId",
            "required": true,
            "example": "554",
            "name": "authorId"
          }
        ]
      }
    },
...
    "/post/comment/secondary/{id}": {
      "get": {
        "summary": "select secondary postComments by id",
        "deprecated": false,
        "description": "select secondary postComments by id",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 522,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentsecondary{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "id",
            "required": true,
            "example": "235",
            "name": "id"
          }
        ]
      }
    },
...
    "/post/{id}": {
      "delete": {
        "summary": "delete post by id",
        "deprecated": false,
        "description": "delete post by id",
        "tags": [
          "post controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 70,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "post{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "id",
            "required": true,
            "example": "822",
            "name": "id"
          }
        ]
      },
      "get": {
        "summary": "select post by id",
        "deprecated": false,
        "description": "select post by id",
        "tags": [
          "post controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 369,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "post{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "id",
            "required": true,
            "example": "397",
            "name": "id"
          }
        ]
      }
    },
...
    "/post/comment/page/author/{authorId}": {
      "get": {
        "summary": "select all postComments by authorId and page",
        "deprecated": false,
        "description": "select all postComments by authorId and page",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 110,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentpageauthor{authorId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "authorId",
            "required": true,
            "example": "931",
            "name": "authorId"
          }
        ]
      }
    },
...
    "/post/author/{authorId}": {
      "get": {
        "summary": "select posts by authorId",
        "deprecated": false,
        "description": "select posts by authorId",
        "tags": [
          "post controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 64,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postauthor{authorId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "authorId",
            "required": true,
            "example": "225",
            "name": "authorId"
          }
        ]
      }
    },
...
    "/post/comment/page/post/{postId}": {
      "get": {
        "summary": "select all postComments by postId and page",
        "deprecated": false,
        "description": "select all postComments by postId and page",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 258,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentpagepost{postId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "  postId",
            "required": true,
            "example": "296",
            "name": "postId"
          }
        ]
      }
    },
...
    "/pin/comment/page/author/{authorId}": {
      "get": {
        "summary": "select all pinComments by authorId and page",
        "deprecated": false,
        "description": "select all pinComments by authorId and page",
        "tags": [
          "pin comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 293,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincommentpageauthor{authorId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "authorId",
            "required": true,
            "example": "938",
            "name": "authorId"
          }
        ]
      }
    },
...
    "/tag/post/{postId}": {
      "get": {
        "summary": "select all tags by postId",
        "deprecated": false,
        "description": "select all tags by postId",
        "tags": [
          "tag controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 780,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "tagpost{postId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "postId",
            "required": true,
            "example": "59",
            "name": "postId"
          }
        ]
      }
    },
...
    "/pin/comment/secondary/{id}": {
      "get": {
        "summary": "select secondary PinComments by id",
        "deprecated": false,
        "description": "select secondary PinComments by id",
        "tags": [
          "pin comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 710,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincommentsecondary{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "id",
            "required": true,
            "example": "754",
            "name": "id"
          }
        ]
      }
    },
...
    "/post/comment/page/secondary/{id}": {
      "get": {
        "summary": "select secondary comments by id and page",
        "deprecated": false,
        "description": "select secondary comments by id and page",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 538,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentpagesecondary{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "      id",
            "required": true,
            "example": "461",
            "name": "id"
          }
        ]
      }
    },
...
    "/post/comment/post/{postId}": {
      "get": {
        "summary": "select all postComments by postId",
        "deprecated": false,
        "description": "select all postComments by postId",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 602,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentpost{postId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "postId",
            "required": true,
            "example": "621",
            "name": "postId"
          }
        ]
      }
    },
...
    "/column/count/post/{columnId}": {
      "get": {
        "summary": "count posts num by column id",
        "deprecated": false,
        "description": "count posts num by column id",
        "tags": [
          "column controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 909,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "columncountpost{columnId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "column id",
            "required": true,
            "example": "673",
            "name": "columnId"
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "okito",
    "version": "1.0.0"
  },
...
}